+++ src/components/GroupedBarChart/types.ts
@@ -1,7 +1,0 @@
-import {Color} from 'types';
-
-export interface Data {
-  data: number[];
-  color: Color;
-  label: string;
-}