+++ src/components/MultiSeriesBarChart/types.ts
@@ -1,0 +1,16 @@
+import {Series} from 'd3-shape';
+import {Color} from 'types';
+
+export interface Data {
+  data: number[];
+  color: Color;
+  highlightColor?: Color;
+  label: string;
+}
+
+export type StackSeries = Series<
+  {
+    [key: string]: number;
+  },
+  string
+>;