+++ src/components/index.ts
@@ -6,7 +6,7 @@ export {NormalizedStackedBar} from './NormalizedStackedBar';
 export {Sparkline} from './Sparkline';
 export {YAxis} from './YAxis';
 export {TooltipContainer} from './TooltipContainer';
-export {GroupedBarChart} from './GroupedBarChart';
 export {SquareColorPreview} from './SquareColorPreview';
 export {StackedAreaChart} from './StackedAreaChart';
 export {LinearXAxis} from './LinearXAxis';
+export {MultiSeriesBarChart} from './MultiSeriesBarChart';